requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
pymysql
python-dotenv
google-generativeai